# Intern label KVs as unique.Handle[string] 