--- a/felix/calc/networkset_lookup_cache.go
+++ b/felix/calc/networkset_lookup_cache.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/projectcalico/calico/felix/dispatcher"
 	"github.com/projectcalico/calico/felix/ip"
-	"github.com/projectcalico/calico/lib/std/internedlabels"
+	"github.com/projectcalico/calico/lib/std/uniquelabels"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/api"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/model"
 	"github.com/projectcalico/calico/libcalico-go/lib/set"
@@ -45,7 +45,7 @@ func init() {
 type networkSetData struct {
 	cidrs  set.Set[ip.CIDR]
 	key    model.NetworkSetKey
-	labels internedlabels.Map
+	labels uniquelabels.Map
 }
 
 func (n networkSetData) IsLocal() bool {
@@ -68,7 +68,7 @@ func (n networkSetData) Key() model.Key {
 	return n.key
 }
 
-func (n networkSetData) Labels() internedlabels.Map {
+func (n networkSetData) Labels() uniquelabels.Map {
 	return n.labels
 }