--- a/felix/collector/utils/utils.go
+++ b/felix/collector/utils/utils.go
@@ -20,6 +20,8 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/projectcalico/calico/lib/std/internedlabels"
+	"github.com/projectcalico/calico/lib/std/unique"
 	"github.com/projectcalico/calico/libcalico-go/lib/backend/model"
 	net2 "github.com/projectcalico/calico/libcalico-go/lib/net"
 )
@@ -80,19 +82,11 @@ func MustParseNet(n string) net2.IPNet {
 	return *cidr
 }
 
-func IntersectLabels(in, out map[string]string) map[string]string {
-	common := map[string]string{}
-	for k := range out {
+func IntersectAndFilterLabels(in, out internedlabels.Map) internedlabels.Map {
+	return out.IntersectAndFilter(in, func(k unique.String, _ unique.String) bool {
 		// Skip Calico labels from the logs
-		if strings.HasPrefix(k, "projectcalico.org/") {
-			continue
-		}
-		if v, ok := in[k]; ok && v == out[k] {
-			common[k] = v
-		}
-	}
-
-	return common
+		return !strings.HasPrefix(k.Value(), "projectcalico.org/")
+	})
 }
 
 // There is support for both global and namespaced networkset. In case of