--- a/felix/labelindex/labelnamevalueindex/label_name_value_index.go
+++ b/felix/labelindex/labelnamevalueindex/label_name_value_index.go
@@ -17,11 +17,10 @@ package labelnamevalueindex
 import (
 	"fmt"
 	"iter"
-	"unique"
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/projectcalico/calico/lib/std/uniquestr"
+	"github.com/projectcalico/calico/lib/std/unique"
 	"github.com/projectcalico/calico/libcalico-go/lib/selector/parser"
 	"github.com/projectcalico/calico/libcalico-go/lib/set"
 )
@@ -40,23 +39,23 @@ import (
 type LabelNameValueIndex[ItemID comparable, Item Labeled] struct {
 	nameOfTrackedItems    string
 	allItems              map[ItemID]Item
-	labelNameToValueToIDs map[unique.Handle[string]]values[ItemID]
+	labelNameToValueToIDs map[unique.String]values[ItemID]
 }
 
 type Labeled interface {
-	OwnLabelHandles() iter.Seq2[unique.Handle[string], unique.Handle[string]]
+	OwnLabelHandles() iter.Seq2[unique.String, unique.String]
 }
 
 func New[ItemID comparable, Item Labeled](nameOfTrackedItems string) *LabelNameValueIndex[ItemID, Item] {
 	return &LabelNameValueIndex[ItemID, Item]{
 		nameOfTrackedItems:    nameOfTrackedItems,
 		allItems:              map[ItemID]Item{},
-		labelNameToValueToIDs: map[unique.Handle[string]]values[ItemID]{},
+		labelNameToValueToIDs: map[unique.String]values[ItemID]{},
 	}
 }
 
 type values[ItemID comparable] struct {
-	m     map[unique.Handle[string]]set.Set[ItemID]
+	m     map[unique.String]set.Set[ItemID]
 	count int
 }
 
@@ -78,7 +77,7 @@ func (idx *LabelNameValueIndex[ItemID, Item]) Add(id ItemID, item Item) {
 		vals, ok := idx.labelNameToValueToIDs[k]
 		if !ok {
 			vals = values[ItemID]{
-				m: map[unique.Handle[string]]set.Set[ItemID]{},
+				m: map[unique.String]set.Set[ItemID]{},
 			}
 			idx.labelNameToValueToIDs[k] = vals
 		}
@@ -119,7 +118,7 @@ func (idx *LabelNameValueIndex[ItemID, Item]) Remove(id ItemID) {
 // label name and selector LabelRestriction (which should be the restriction
 // for that label).  If the LabelRestriction is not "useful", returns
 // FullScanStrategy().
-func (idx *LabelNameValueIndex[ItemID, Item]) StrategyFor(labelName unique.Handle[string], r parser.LabelRestriction) ScanStrategy[ItemID] {
+func (idx *LabelNameValueIndex[ItemID, Item]) StrategyFor(labelName unique.String, r parser.LabelRestriction) ScanStrategy[ItemID] {
 	if !r.MustBePresent {
 		// Not much we can do if the selector doesn't match on this label.
 		return FullScanStrategy[ItemID, Item]{allItems: idx.allItems}
@@ -140,7 +139,7 @@ func (idx *LabelNameValueIndex[ItemID, Item]) StrategyFor(labelName unique.Handl
 	// If we get here, then the selector does match on this label, and it cares
 	// about specific values. Whittle down the list of values to the ones that
 	// match objects that we're tracking.
-	var filteredMustHaves []unique.Handle[string]
+	var filteredMustHaves []unique.String
 	var idSets []set.Set[ItemID]
 	for _, v := range r.MustHaveOneOfValues {
 		if idsSet := idx.labelNameToValueToIDs[labelName].m[v]; idsSet != nil {
@@ -153,7 +152,7 @@ func (idx *LabelNameValueIndex[ItemID, Item]) StrategyFor(labelName unique.Handl
 		// We filtered all values out!  That means that the selector cannot
 		// match anything.  If it could match something, we'd have found it
 		// in the index.
-		logrus.Debugf("No %s with %s=%v", idx.nameOfTrackedItems, labelName.Value(), uniquestr.SliceStringer(r.MustHaveOneOfValues))
+		logrus.Debugf("No %s with %s=%v", idx.nameOfTrackedItems, labelName.Value(), unique.SliceStringer(r.MustHaveOneOfValues))
 		return NoMatchStrategy[ItemID]{}
 	}
 
@@ -236,8 +235,8 @@ func (n NoMatchStrategy[ItemID]) Name() string {
 // a specific value for a certain label.  It is the narrowest, most optimized
 // strategy.
 type LabelNameSingleValueStrategy[ItemID comparable] struct {
-	label unique.Handle[string]
-	value unique.Handle[string]
+	label unique.String
+	value unique.String
 	idSet set.Set[ItemID]
 }
 
@@ -266,13 +265,13 @@ func (s LabelNameSingleValueStrategy[ItemID]) Name() string {
 // LabelNameMultiValueStrategy is a ScanStrategy that scans over items that have
 // specific, values for a certain label.
 type LabelNameMultiValueStrategy[ItemID comparable] struct {
-	label  unique.Handle[string]
-	values []unique.Handle[string]
+	label  unique.String
+	values []unique.String
 	idSets []set.Set[ItemID]
 }
 
 func (s LabelNameMultiValueStrategy[ItemID]) String() string {
-	return fmt.Sprintf("scan multi label %s=%v", s.label.Value(), uniquestr.SliceStringer(s.values))
+	return fmt.Sprintf("scan multi label %s=%v", s.label.Value(), unique.SliceStringer(s.values))
 }
 
 func (s LabelNameMultiValueStrategy[ItemID]) EstimatedItemsToScan() int {
@@ -297,7 +296,7 @@ func (s LabelNameMultiValueStrategy[ItemID]) Name() string {
 // particular label, no matter the value of that label.  It is used for
 // selectors such as "has(labelName)".
 type LabelNameStrategy[ItemID comparable] struct {
-	label  unique.Handle[string]
+	label  unique.String
 	values values[ItemID]
 }